Added snackbar when deleting categories

+++ app/src/main/java/io/connorwyatt/flashcards/activities/CategoriesActivity.kt
@@ -3,6 +3,8 @@ package io.connorwyatt.flashcards.activities
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
+import android.support.design.widget.CoordinatorLayout
+import android.support.design.widget.Snackbar
 import android.support.v7.app.AppCompatActivity
 import android.support.v7.widget.LinearLayoutManager
 import android.support.v7.widget.RecyclerView
@@ -15,33 +17,46 @@ import io.connorwyatt.flashcards.services.CategoryService
 class CategoriesActivity : AppCompatActivity()
 {
     private val categoryService = CategoryService(this)
-    private var categoryItems: List<CategoryListAdapter.ListItem>? = null
+    private var categoryItems: MutableList<CategoryListAdapter.ListItem> = mutableListOf()
+    private val removedCategoryIds: MutableList<Long> = mutableListOf()
 
     override fun onCreate(savedInstanceState: Bundle?)
     {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_categories)
 
-        categoryItems = getCategoryListItems()
+        categoryItems = getCategoryListItems().toMutableList()
 
         setUpToolbar()
 
-        setUpRecycler(categoryItems as List<CategoryListAdapter.ListItem>)
+        setUpRecycler(categoryItems.toList())
     }
 
     private fun getCategoryListItems(): List<CategoryListAdapter.ListItem>
     {
         val categories = categoryService.getAll()
 
-        val categoryListItems: List<CategoryListAdapter.ListItem> = categories.map(fun(category: Category): CategoryListAdapter.ListItem
-        {
-            val flashcardCountForCategory = categoryService.getFlashcardsForCategory(category.id).size
-            val averageRating = categoryService.getAverageRatingForCategory(category.id)
+        val categoryListItems: List<CategoryListAdapter.ListItem> =
+            categories.map(
+                fun(category: Category): CategoryListAdapter.ListItem
+                {
+                    val flashcardCountForCategory =
+                        categoryService.getFlashcardsForCategory(category.id).size
+                    val averageRating = categoryService.getAverageRatingForCategory(category.id)
 
-            return CategoryListAdapter.ListItem(category, flashcardCountForCategory, averageRating)
-        })
+                    return CategoryListAdapter.ListItem(
+                        category,
+                        flashcardCountForCategory,
+                        averageRating)
+                }
+            )
 
         return categoryListItems
+    }
+
+    private fun getFilteredCategoryListItems(): List<CategoryListAdapter.ListItem>
+    {
+        return categoryItems.filterNot { categoryItem -> removedCategoryIds.contains(categoryItem.category.id) }
     }
 
     private fun setUpToolbar()
@@ -53,12 +68,38 @@ class CategoriesActivity : AppCompatActivity()
 
     private fun setUpRecycler(categoryListItems: List<CategoryListAdapter.ListItem>)
     {
+        val coordinatorLayout = findViewById(R.id.categories_coordinator_layout) as CoordinatorLayout
         val categoryListAdapter = CategoryListAdapter(categoryListItems)
 
         categoryListAdapter.addOnDeleteListener { category ->
-            categoryService.delete(category)
+            removedCategoryIds.add(category.id)
 
-            categoryListAdapter.updateData(getCategoryListItems())
+            categoryListAdapter.updateData(getFilteredCategoryListItems())
+
+            val snackbar = Snackbar.make(coordinatorLayout,
+                                         getString(R.string.deleted_category_snackbar,
+                                                   category.name),
+                                         Snackbar.LENGTH_LONG)
+                .setAction(getString(R.string.action_undo), { view ->
+                    removedCategoryIds.remove(category.id)
+                    categoryListAdapter.updateData(getFilteredCategoryListItems())
+                })
+                .setCallback(object : Snackbar.Callback()
+                             {
+                                 override fun onDismissed(snackbar: Snackbar?, event: Int)
+                                 {
+                                     if (removedCategoryIds.contains(category.id))
+                                     {
+                                         removedCategoryIds.remove(category.id)
+                                         categoryService.delete(category)
+                                         categoryItems.removeAll { categoryItem -> categoryItem.category.id === category.id }
+                                         categoryListAdapter.updateData(
+                                             getFilteredCategoryListItems())
+                                     }
+                                 }
+                             })
+
+            snackbar.show()
         }
 
         val recycler = findViewById(R.id.categories_recycler) as RecyclerView
